Use hash table internally in `group-by`, and add more stability guarantees.
This avoids the execution time depending on the number of equivalence classes,
but only in cases where the equality predicate corresponds to that used by a
kind of hash table.

Significantly faster, except in cases where the number of classes is small.
Also calls the key function less often, which can help when that function is
expensive.